bug fixes & rename
